fixed error in Calc_Site_MeanLength
oops.. I just noticed that Calc_Site_MeanLength() didn’t use the
min_size parameter passed into it - because of some commented out code.
I think the comment out was because of failure for SIZE_ of NA, and so
have put the code back in, with a fix for NAs in the field.