Minor fix to site level data
The previous fix (LAY-19 -> LIS-19) was incorrect. The site labelled
LAY-19 was from a survey at Lisianski at a fixed EAR site. Changed the
site name to LIS-EAR, exclude flag to -1 and the type to FIXED. Still
need to update the fish obs file once Troy makes the correction in
Oracle.