Fixed Benthos not summing to 100%
Went back through the datasheets for replicates where visually
estimated benthic cover did not sum to 100%. I have corrected most of
them at NCRMP sites. In a few cases, it was not obvious what the fix
was and I couldn’t find one sheet. Thus there are still around 6
replicates with values in range of 90-100%. Long term, we should
probably standardize that data to 100%, but leaving it for now.