Added Calc_Pooled_Simple()
The new function is intended to greatly simplify the process to turn
strata data into pooled up island or regional data (or whatever). It
makes fewer assumption about the structure of the data and does not
require a link to the sectors table. Also, is capable of pooling for
multiple sample designs (in different years) in one call. If using this
function, onus is on programmer to ensure that all the relevant pooling
fields (ANALYSIS_SEC, ANALYSIS_YEAR) are correctly specified before the
data are pooled to strata level.