Updated Munge to deal with new SIteMaster file
SiteMaster file can now have multiple surveys at the same SITE (e.g. a
fish and a benthic survey at a SITE). Those are uniquely identified on
SITEVISITID, so it would be possible to use that to join fish survey
and site master records. However, for here, I have gone for the simpler
patch - which is to filter the site master records to just fish survey
records. Among fish surveys, it is possible to treat SITE as a unique
identifier.