Dealing with duplicate SITE surveys
Now we can have fish and benthic surveys at the same site - and therefore can be 2 entries for a given SITE value in the SITEMASTER table. Thus, the merge between raw fish data and the site master table wouldbe problemattic if done on SITE.

I have changed the code to merge on SITE and METHOD. Should not be fine. We might consider instead merging on SITEVISITID  (which should be unique to the SITE and METHOD combination) - but holding off on that for now