Updated SURVEY Master with missing depth data
Tye pulled fish depth data and integrated it into the SURVEY MASTER file as new_depth_min and max. Courtney manually reviewed the data and merged together the 6 depth columns. The original SURVEY MASTER pre merge is also in this folder. We still have issues with some of the depths not matching the depth bin.